Embedded LND: Advanced settings
