+++ locales/en.json
@@ -97,6 +97,7 @@
     "general.reorder": "Reorder",
     "general.yes": "Yes",
     "general.no": "No",
+    "general.advanced": "Advanced",
     "restart.title": "Restart required",
     "restart.msg": "ZEUS has to be restarted before the new configuration is applied.",
     "restart.msg1": "Would you like to restart now?",
