+++ views/Settings/EmbeddedNode/index.tsx
@@ -1,24 +1,14 @@
 import * as React from 'react';
-import {
-    Alert,
-    NativeModules,
-    Platform,
-    ScrollView,
-    Text,
-    View
-} from 'react-native';
+import { ScrollView, View } from 'react-native';
 import { Icon, ListItem } from 'react-native-elements';
 import { inject, observer } from 'mobx-react';
 
-import Button from '../../../components/Button';
 import Header from '../../../components/Header';
 import Screen from '../../../components/Screen';
-import Switch from '../../../components/Switch';
 
 import SettingsStore from '../../../stores/SettingsStore';
 
 import { localeString } from '../../../utils/LocaleUtils';
-import { sleep } from '../../../utils/SleepUtils';
 import { themeColor } from '../../../utils/ThemeUtils';
 
 interface EmbeddedNodeProps {
@@ -26,51 +16,12 @@ interface EmbeddedNodeProps {
     SettingsStore: SettingsStore;
 }
 
-interface EmbeddedNodeState {
-    rescan: boolean | undefined;
-    compactDb: boolean | undefined;
-}
-
-const restartNeeded = () => {
-    const title = localeString('restart.title');
-    const message = localeString('restart.msg');
-    if (Platform.OS === 'android') {
-        Alert.alert(title, message + '\n' + localeString('restart.msg1'), [
-            {
-                style: 'cancel',
-                text: localeString('general.no')
-            },
-            {
-                style: 'default',
-                text: localeString('general.yes'),
-                onPress: async () => {
-                    try {
-                        // await NativeModules.ZeusTor.stopTor();
-                        await NativeModules.LndMobile.stopLnd();
-                        await NativeModules.LndMobileTools.killLnd();
-                    } catch (e) {
-                        console.log(e);
-                    }
-                    NativeModules.LndMobileTools.restartApp();
-                }
-            }
-        ]);
-    } else {
-        Alert.alert(title, message);
-    }
-};
-
 @inject('SettingsStore')
 @observer
 export default class EmbeddedNode extends React.Component<
     EmbeddedNodeProps,
-    EmbeddedNodeState
+    {}
 > {
-    state = {
-        rescan: false,
-        compactDb: false
-    };
-
     async UNSAFE_componentWillMount() {
         const { SettingsStore } = this.props;
         const { settings } = SettingsStore;
@@ -83,14 +34,8 @@ export default class EmbeddedNode extends React.Component<
 
     render() {
         const { navigation, SettingsStore } = this.props;
-        const { rescan, compactDb } = this.state;
-        const { updateSettings, embeddedLndNetwork, settings }: any =
-            SettingsStore;
-        const {
-            automaticDisasterRecoveryBackup,
-            bimodalPathfinding,
-            expressGraphSync
-        } = settings;
+        const { embeddedLndNetwork, settings }: any = SettingsStore;
+        const { automaticDisasterRecoveryBackup, expressGraphSync } = settings;
 
         return (
             <Screen>
@@ -195,43 +140,6 @@ export default class EmbeddedNode extends React.Component<
                             containerStyle={{
                                 backgroundColor: 'transparent'
                             }}
-                            onPress={() => navigation.navigate('Pathfinding')}
-                        >
-                            <ListItem.Content>
-                                <ListItem.Title
-                                    style={{
-                                        color: themeColor('text'),
-                                        fontFamily: 'PPNeueMontreal-Book'
-                                    }}
-                                >
-                                    {localeString(
-                                        'views.Settings.EmbeddedNode.Pathfinding.title'
-                                    )}
-                                </ListItem.Title>
-                                <ListItem.Title
-                                    style={{
-                                        color: themeColor('secondaryText'),
-                                        fontFamily: 'PPNeueMontreal-Book'
-                                    }}
-                                >
-                                    {bimodalPathfinding
-                                        ? localeString(
-                                              'views.Settings.EmbeddedNode.bimodal'
-                                          )
-                                        : localeString(
-                                              'views.Settings.EmbeddedNode.apriori'
-                                          )}
-                                </ListItem.Title>
-                            </ListItem.Content>
-                            <Icon
-                                name="keyboard-arrow-right"
-                                color={themeColor('secondaryText')}
-                            />
-                        </ListItem>
-                        <ListItem
-                            containerStyle={{
-                                backgroundColor: 'transparent'
-                            }}
                             onPress={() => navigation.navigate('Peers')}
                         >
                             <ListItem.Content>
@@ -275,146 +183,31 @@ export default class EmbeddedNode extends React.Component<
                                 color={themeColor('secondaryText')}
                             />
                         </ListItem>
-                        <>
-                            <ListItem
-                                containerStyle={{
-                                    borderBottomWidth: 0,
-                                    backgroundColor: 'transparent'
-                                }}
-                            >
+                        <ListItem
+                            containerStyle={{
+                                backgroundColor: 'transparent'
+                            }}
+                            onPress={() =>
+                                navigation.navigate(
+                                    'EmbeddedNodeSettingsAdvanced'
+                                )
+                            }
+                        >
+                            <ListItem.Content>
                                 <ListItem.Title
                                     style={{
                                         color: themeColor('text'),
                                         fontFamily: 'PPNeueMontreal-Book'
                                     }}
                                 >
-                                    {localeString(
-                                        'views.Settings.EmbeddedNode.rescan'
-                                    )}
+                                    {localeString('general.advanced')}
                                 </ListItem.Title>
-                                <View
-                                    style={{
-                                        flex: 1,
-                                        flexDirection: 'row',
-                                        justifyContent: 'flex-end'
-                                    }}
-                                >
-                                    <Switch
-                                        value={rescan}
-                                        onValueChange={async () => {
-                                            this.setState({
-                                                rescan: !rescan
-                                            });
-                                            await updateSettings({
-                                                rescan: !rescan
-                                            });
-                                        }}
-                                    />
-                                </View>
-                            </ListItem>
-                            <View
-                                style={{
-                                    margin: 10
-                                }}
-                            >
-                                <Text
-                                    style={{
-                                        color: themeColor('secondaryText')
-                                    }}
-                                >
-                                    {localeString(
-                                        'views.Settings.EmbeddedNode.rescan.subtitle'
-                                    )}
-                                </Text>
-                            </View>
-                        </>
-                        <>
-                            <ListItem
-                                containerStyle={{
-                                    borderBottomWidth: 0,
-                                    backgroundColor: 'transparent'
-                                }}
-                            >
-                                <ListItem.Title
-                                    style={{
-                                        color: themeColor('text'),
-                                        fontFamily: 'PPNeueMontreal-Book'
-                                    }}
-                                >
-                                    {localeString(
-                                        'views.Settings.EmbeddedNode.compactDb'
-                                    )}
-                                </ListItem.Title>
-                                <View
-                                    style={{
-                                        flex: 1,
-                                        flexDirection: 'row',
-                                        justifyContent: 'flex-end'
-                                    }}
-                                >
-                                    <Switch
-                                        value={compactDb}
-                                        onValueChange={async () => {
-                                            this.setState({
-                                                compactDb: !compactDb
-                                            });
-                                            await updateSettings({
-                                                compactDb: !compactDb
-                                            });
-                                        }}
-                                    />
-                                </View>
-                            </ListItem>
-                            <View
-                                style={{
-                                    margin: 10
-                                }}
-                            >
-                                <Text
-                                    style={{
-                                        color: themeColor('secondaryText')
-                                    }}
-                                >
-                                    {localeString(
-                                        'views.Settings.EmbeddedNode.compactDb.subtitle'
-                                    )}
-                                </Text>
-                            </View>
-                        </>
-                        <>
-                            <View style={{ margin: 10 }}>
-                                <Button
-                                    title={localeString(
-                                        'views.Settings.EmbeddedNode.stopLndDeleteNeutrino'
-                                    )}
-                                    onPress={async () => {
-                                        try {
-                                            await NativeModules.LndMobile.stopLnd();
-                                            await sleep(5000); // Let lnd close down
-                                        } catch (e: any) {
-                                            // If lnd was closed down already
-                                            if (
-                                                e?.message?.includes?.('closed')
-                                            ) {
-                                                console.log('yes closed');
-                                            } else {
-                                                console.error(e.message, 10000);
-                                                return;
-                                            }
-                                        }
-
-                                        console.log(
-                                            await NativeModules.LndMobileTools.DEBUG_deleteNeutrinoFiles(
-                                                embeddedLndNetwork === 'Mainnet'
-                                                    ? 'mainnet'
-                                                    : 'testnet'
-                                            )
-                                        );
-                                        restartNeeded();
-                                    }}
-                                />
-                            </View>
-                        </>
+                            </ListItem.Content>
+                            <Icon
+                                name="keyboard-arrow-right"
+                                color={themeColor('secondaryText')}
+                            />
+                        </ListItem>
                     </ScrollView>
                 </View>
             </Screen>
